requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
